wip refactor custom hook
