+++ src/hooks/index.type.ts
@@ -1,0 +1,4 @@
+export type UseSignupProps = {
+  email: string;
+  password: string;
+};
