+++ src/pages/login.tsx
@@ -1,94 +1,19 @@
 import { Header } from "@/components/Header";
-import {
-  onAuthStateChanged,
-  signInWithEmailAndPassword,
-  signInWithRedirect,
-} from "@firebase/auth";
-import { auth, db, provider } from "../../lib/firebase";
-import { useRouter } from "next/router";
-import { useAuthContext } from "@/context/AuthContext";
-import { collection, doc, getDoc, getDocs, setDoc } from "firebase/firestore";
-import { useEffect, useState } from "react";
 import Link from "next/link";
 import GoogleIcon from "@mui/icons-material/Google";
 import { Meta } from "@/components/Meta";
+import { useLogin } from "@/hooks/useLogin";
 
 export default function Login() {
-  const router = useRouter();
-  const { currentUser } = useAuthContext(); //ログイン状態
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-
-  useEffect(() => {
-    onAuthStateChanged(auth, (user) => {
-      user && router.push("/");
-    });
-  }, []);
-
-  /**
-   * Googleアカウントでログイン
-   */
-  const onGoogleLogin = async (e: any) => {
-    e.preventDefault();
-    await signInWithRedirect(auth, provider);
-  };
-
-  /**
-   * メールアドレスのログイン
-   */
-  const onEmailLogin = async (e: any) => {
-    e.preventDefault();
-    try {
-      signInWithEmailAndPassword(auth, email, password)
-        .then(async (userCredential) => {
-          const user = userCredential.user;
-
-          const docRef = doc(db, "user", user.uid);
-          const docSnap = await getDoc(docRef);
-          if (!docSnap.exists()) {
-            await setDoc(doc(db, "user", user.uid), {});
-          }
-          router.push("/");
-        })
-        .catch((error) => {
-          console.log(error);
-          alert("ログインに失敗しました");
-        });
-    } catch {
-      alert("ログインエラーが起きました");
-    }
-  };
-
-  /**
-   * ゲスト用ののログイン
-   */
-  const onGuestLogin = async (e: any) => {
-    // setIsAllLoad(true);
-    try {
-      signInWithEmailAndPassword(
-        auth,
-        process.env.NEXT_PUBLIC_GUEST_EMAIL!,
-        process.env.NEXT_PUBLIC_GUEST_PASSWORD!
-      )
-        .then(async (userCredential) => {
-          const user = userCredential.user;
-
-          const docRef = doc(db, "user", user.uid);
-          const docSnap = await getDoc(docRef);
-          if (!docSnap.exists()) {
-            await setDoc(doc(db, "user", user.uid), {});
-          }
-          // setIsAllLoad(false);
-          router.push("/");
-        })
-        .catch((error) => {
-          console.log(error);
-          alert("ログインに失敗しました");
-        });
-    } catch {
-      alert("ログインエラーが起きました");
-    }
-  };
+  const {
+    email,
+    setEmail,
+    password,
+    setPassword,
+    onGoogleLogin,
+    onEmailLogin,
+    onGuestLogin,
+  } = useLogin();
 
   return (
     <>
